Separate metrics from controller to used it from outside
